% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{buildBipartiteGraph}
\alias{buildBipartiteGraph}
\title{build Bipartite Graph}
\usage{
buildBipartiteGraph(object.list, features.integration, assay = NULL)
}
\arguments{
\item{object.list}{List of Seurat objects}

\item{features.integration}{feature to use}

\item{assay}{assay to use}
}
\value{
a Bipartite Graph
}
\description{
build Bipartite Graph
}
