% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_points.R
\name{chebyshev_nodes}
\alias{chebyshev_nodes}
\title{Chebyshev Nodes}
\usage{
chebyshev_nodes(n, interval = c(0, 1))
}
\arguments{
\item{n}{The number of points to generate}

\item{interval}{The interval over which to calculate Chebyshev nodes}
}
\value{
A numeric vector containing n Chebyshev nodes in the interval 'interval'
}
\description{
Calculate Chebyshev points for interpolation
}
