% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{computeProportion}
\alias{computeProportion}
\title{Stacked bar plot showing the proportion of cells across certrain cell groups}
\usage{
computeProportion(
  object,
  x = "cellType",
  fill = "condition",
  facet = NULL,
  colors.use = NULL,
  n.colors = 8,
  n.row = 1,
  title.name = NULL,
  legend.title = NULL,
  xlabel = NULL,
  ylabel = "Cellular composition (\%)",
  width = 0.8,
  show.legend = TRUE,
  x.lab.rot = TRUE,
  text.size = 10,
  flip = FALSE
)
}
\arguments{
\item{object}{Seurat object}

\item{x}{Name of one metadata column to show on the x-axis}

\item{fill}{Name of one metadata column to compare the proportion of cells}

\item{facet}{Name of one metadata column defining faceting groups}

\item{colors.use}{defining the color of stacked bar plot; either a char vector defining a color for each cell group or a palette name from brewer.pal}

\item{n.colors}{Number of colors when setting colors.use to be a palette name from brewer.pal}

\item{n.row}{Number of rows in facet_grid()}

\item{title.name}{Name of the main title}

\item{legend.title}{Name of legend}

\item{xlabel}{Name of x label}

\item{ylabel}{Name of y label}

\item{width}{bar width}

\item{show.legend}{Whether show the legend}

\item{x.lab.rot}{Whether rorate the xtick labels}

\item{text.size}{font size}

\item{flip}{Whether flip the cartesian coordinates so that horizontal becomes vertical}
}
\value{
ggplot2 object
}
\description{
Stacked bar plot showing the proportion of cells across certrain cell groups
}
