% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula.R
\name{copula_sampler}
\alias{copula_sampler}
\title{Generate a sample from a copula}
\usage{
copula_sampler(copula, N, gss = 1.37)
}
\arguments{
\item{copula}{An R object of class "Copula"}

\item{N}{An integer indicating the number of collocation points to use for each dimension}
}
\value{
A function which takes one parameter ('n') and returns a matrix with dimensions (n,
  dim(copula)) containing the sample from the copula, each row being one observation.
}
\description{
Generates a sample from the specified copula object
}
