% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_points.R
\name{gaussian_nodes}
\alias{gaussian_nodes}
\title{Gauss-Hermite quadrature nodes}
\usage{
gaussian_nodes(n)
}
\arguments{
\item{n}{The number of points to generate}
}
\value{
A numeric vector containing n Gaussian quadrature nodes
}
\description{
Calculate Gaussian quadrature nodes for interpolation with regard to the
standard normal density as the weighting function
}
\details{
This is a wrapper for optimal_points(n, norm_moment_fun).
}
