% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{identifyClusters_internal}
\alias{identifyClusters_internal}
\title{identify cell clusters of a single Seurat object}
\usage{
identifyClusters_internal(
  object,
  resolution = NULL,
  method = "matrix",
  algorithm = 4,
  resRange = NULL,
  nDims.consensus = 30,
  clustering.method = c("hierarchical", "community"),
  graph.name = NULL,
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{resolution}{the resolution in Leiden algorithm; if it is NULL, the optimal resoultion will be inferred based on eigen spectrum}

\item{method}{Method for running leiden (defaults to matrix which is fast for small datasets). Enable method = "igraph" to avoid casting large data to a dense matrix.}

\item{algorithm}{Algorithm for modularity optimization (1 = original Louvain algorithm; 2 = Louvain algorithm with multilevel refinement; 3 = SLM algorithm; 4 = Leiden algorithm). Leiden requires the leidenalg python.}

\item{resRange}{the range of resolutions in Leiden algorithm; if it is NULL, the optimal range of resoultion will be from 0.1 to 0.5}

\item{nDims.consensus}{the number of singular values to estimate from the consensus matrix.}

\item{clustering.method}{method for performing clustering on the consensus matrix from a range of resolutions}

\item{graph.name}{Name of graph to use for the clustering algorithm}

\item{...}{other parameter passing to FindClusters function in Seurat package}
}
\value{
Seurat object
}
\description{
identify cell clusters of a single Seurat object
}
