% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{identifyConfidentCells}
\alias{identifyConfidentCells}
\title{identify confident cells in each cell cluster}
\usage{
identifyConfidentCells(
  object.list,
  features.integration,
  quantile.cutoff = 0.75,
  min.cluster = 20,
  assay = NULL
)
}
\arguments{
\item{object.list}{a list of Seurat object}

\item{features.integration}{features used to identify confident cells}

\item{quantile.cutoff}{quantile cutoff (default = 0.75)}

\item{min.cluster}{the minimum number of cells in the cluster for identifying confident cells}

\item{assay}{Assay to use}
}
\value{
a list of Seurat object with identified confident cells in each dataset
}
\description{
identify confident cells in each cell cluster
}
