% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{identifyIntegrationFeatures}
\alias{identifyIntegrationFeatures}
\title{Identify integration features}
\usage{
identifyIntegrationFeatures(
  object.list,
  integrationFeatures.method = c("joint", "individual"),
  selection.method = c("vst", "mean.var.plot"),
  nfeatures = 2000,
  mean.cutoff = c(0.01, 5),
  dispersion.cutoff = c(0.25, Inf),
  ...
)
}
\arguments{
\item{object.list}{a lsit of Seurat objects}

\item{integrationFeatures.method}{"joint" or "individual";
"joint": Identify integration features from the concatenated data matrix;
"individual": ranks features by the number of datasets they are deemed variable in, breaking ties by the median variable feature rank across datasets. It returns
the top scoring features by this ranking.}

\item{selection.method}{The method to choose top variable features:
vst: First, fits a line to the relationship of log(variance) and log(mean) using local polynomial regression (loess). Then standardizes the feature values using the observed mean and expected variance (given by the fitted line). Feature variance is then calculated on the standardized values after clipping to a maximum (see clip.max parameter).
mean.var.plot (mvp): First, uses a function to calculate average expression (mean.function) and dispersion (dispersion.function) for each feature. Next, divides features into num.bin (deafult 20) bins based on their average expression, and calculates z-scores for dispersion within each bin. The purpose of this is to identify variable features while controlling for the strong relationship between variability and average expression.}

\item{nfeatures}{Number of features to select as top variable features; only used when selection.method is set to 'vst'}

\item{mean.cutoff}{A two-length numeric vector with low- and high-cutoffs for feature means}

\item{dispersion.cutoff}{A two-length numeric vector with low- and high-cutoffs for feature dispersions}

\item{...}{other parameters in FindVariableFeatures or SelectIntegrationFeatures}
}
\value{
a char vector containing the features for integration analysis
}
\description{
Identify integration features
}
