% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{identifyMarkers}
\alias{identifyMarkers}
\title{Identify marker genes associated with cell clusters}
\usage{
identifyMarkers(
  object,
  mode = c("separate", "integrated"),
  features = NULL,
  test.use = "wilcox",
  only.pos = TRUE,
  min.pct = 0.25,
  logfc.threshold = 0.25,
  ...
)
}
\arguments{
\item{object}{a lsit of Seurat objects or a single Seurat object}

\item{mode}{"separate" or "integrated"; "separate" means identifying marker genes of cell clusters from each dataset separately, and "integrated" means identifying marker genes of cell clusters from the integrated data}

\item{features}{features used to perform statistical test}

\item{test.use}{which test to use}

\item{only.pos}{Only return positive markers}

\item{min.pct}{Threshold of the percent of cells enriched in one cluster}

\item{logfc.threshold}{Threshold of Log Fold Change}

\item{...}{other parameters in FindAllMarkers}
}
\description{
Identify marker genes associated with cell clusters
}
