% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{identifyNeighbors}
\alias{identifyNeighbors}
\title{identify shared nearest neighbors}
\usage{
identifyNeighbors(
  object,
  mode = c("separate", "integrated"),
  nDims.pca = 40,
  force.pca = TRUE,
  nDims.knn = 40,
  k.param = 20,
  prune.SNN = 1/15,
  ...
)
}
\arguments{
\item{object}{a lsit of Seurat objects or a single Seurat object}

\item{mode}{"separate" or "integrate"}

\item{nDims.pca}{the number of dimensions to use for running PCA}

\item{force.pca}{Set force.pca = FALSE to skip the PCA calculation. Default = TRUE will calculate PCA.}

\item{nDims.knn}{the number of dimensions to use for building SNN}

\item{k.param}{Defines k for the k-nearest neighbor algorithm}

\item{prune.SNN}{Sets the cutoff for acceptable Jaccard index when computing the neighborhood overlap for the SNN construction. Any edges with values less than or equal to this will be set to 0 and removed from the SNN graph. Essentially sets the strigency of pruning (0 — no pruning, 1 — prune everything).}

\item{...}{other parameters in FindNeighbors}
}
\value{
seurat object
}
\description{
identify shared nearest neighbors
}
