% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{integrateData}
\alias{integrateData}
\title{Infer a shared embedding of cells across all datasets after removing technical variation}
\usage{
integrateData(
  object,
  mat,
  assay = NULL,
  new.assay.name = NULL,
  nDims.scMC = 40,
  lambda = 1
)
}
\arguments{
\item{object}{unintegrated Seurat object}

\item{mat}{structured matrices obtained from learnTechnicalVariation}

\item{assay}{assay to store the integrated data}

\item{new.assay.name}{Name for the new assay containing the integrated data}

\item{nDims.scMC}{number of dimensions to compute in the scMC integrated space}

\item{lambda}{the tuning parameter, non-negative.}
}
\value{
Seurat object with an integrated assay
}
\description{
Infer a shared embedding of cells across all datasets after removing technical variation
}
