% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{learnTechnicalVariation}
\alias{learnTechnicalVariation}
\title{Learn technical variation between any two datasets}
\usage{
learnTechnicalVariation(
  object.list,
  features.integration,
  similarity.cutoff = 0.6,
  assay = NULL
)
}
\arguments{
\item{object.list}{List of Seurat objects}

\item{features.integration}{feature}

\item{similarity.cutoff}{a thresholding parameter determining whether cell clusters are shared across different datasets based on their similarity. If T is too small, the biological variation may be removed. If T is too large, the technical variation could not be removed.}

\item{assay}{assay to use}
}
\value{
a list of two structured matrices: one is the data matrix and the other is the confounding matrix
}
\description{
Learn technical variation between any two datasets
}
