% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_points.R
\name{norm_moment_fun}
\alias{norm_moment_fun}
\alias{exp_moment_fun}
\alias{unif_moment_fun}
\title{Theoretical moments calculation}
\usage{
norm_moment_fun(k, sd = 1)

exp_moment_fun(k, rate = 1)

unif_moment_fun(k, supp = c(0, 1))
}
\arguments{
\item{k}{The moment which to calculate}

\item{sd}{The standard deviation of the normal distibution (mean is assumed 0)}

\item{rate}{The rate parameter of the exponential distribution}

\item{supp}{The vector (a, b) indicating the support of the uniform distribution}
}
\value{
The k-th moment of the distribution.
}
\description{
Theoretical moments calculation
}
