% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_points.R
\name{normal_moment_fun}
\alias{normal_moment_fun}
\title{Central normal moments calculation}
\usage{
normal_moment_fun(k, sd = 1)
}
\arguments{
\item{k}{The moment which to calculate}

\item{sd}{The standard deviation of the normal distibution (mean is assumed 0)}
}
\value{
The k-th moment of Normal(0, sd) distribution.
}
\description{
Central normal moments calculation
}
