% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_points.R
\name{optimal_points}
\alias{optimal_points}
\title{Calculate optimal collocation points from (sample) moments of a distribution}
\usage{
optimal_points(N, moment_fun = NULL, sample = NULL)
}
\arguments{
\item{N}{The number of collocation points to calculate}

\item{moment_fun}{A function which takes one (unsigned) integer k and return the k-th moment of
the distribution to use for collocation. Only 2*N moments are needed.}

\item{sample}{If moment_fun not specified, the sample from which to calculate the collocation
points using the 2*N sample moments.}
}
\value{
The N optimal collocation points for interpolation used in SCMC.
}
\description{
Calculates optimal collocation points from (sample) moments of the approximating distribution using the Golub-Welsch algorithm
}
\details{
Note: Using sample moments for a large sample from the approximating random variable gives
similar results as using theoretical. For example, using sample = rnorm(1e6) gives similar points
as using moment_fun = scmc::normal_moment_fun.
}
