% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmc.R
\docType{package}
\name{scmc}
\alias{scmc}
\alias{scmc-package}
\title{scmc: Univariate and multivariate sampling using Stochastic Collocation Monte Carlo.}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
}
\author{
Blagoje Ivanovic <blagoje.r.ivanovic@gmail.com>
}
