% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_points.R
\name{norm_moment_fun}
\alias{norm_moment_fun}
\title{Theoretical moments calculation}
\usage{
norm_moment_fun(k, sd = 1)
}
\arguments{
\item{k}{The moment which to calculate}

\item{sd}{The standard deviation of the normal distibution (mean is assumed 0)}
}
\value{
The k-th moment of the distribution.
}
\description{
Theoretical moments calculation
}
